Match any version number in popolo filename
Previously the regex only matched version 1.0, so when a new version
came out this would silently fail because the filename wouldn't be
matched. This change means we should catch new versions of the file. If
there's a problem with the format of the file then at least the program
will error noisily.